requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pillow
psycopg2
PyJWT
sqlparse
typing_extensions
tzdata
